feat: support log files and journald
Refactors how path arguments are handled in the CLI to lower
maintenance burden of adding new paths in the future compared
to the old system, and adds support for log files and journald.

The filter is controlled separately for files/journald and
stdout - only stdout responds to `RUST_LOG` etc.

To set the filter for the log file or journald, you can now
pass `--log.filter <FILTER>`