refactor(stages/execution): make ExecutionStage own an Executor
before, we were only storing a chainspec and not an executor, which resulted
in a weird API with functions with a lot of arguments..this refactors the stage
to be able to be constructed from an executor or a chainspec.

this in turn will allow us to instantiate a stage with an executor that
e.g. is loaded with an inspector and have it be used across all transactions