feat: database metrics
Adds two database-specific metrics:

- `db.table_size` which is the size of a table (in bytes)
- `db.table_pages` which is the number of pages in a table

Both are labelled by the table (`table`), and the pages metric
is also labelled by the page type (`type`).

This also refactors how the Prometheus endpoint and recorder
are installed, allowing for a hook that manually scrapes
metrics for components that cannot push metric updates live.

This hook is called every time the metric endpoint is served.