fix: storage wipe changeset
The storage changeset cursor moves through all
storage slots if the storage was wiped during a block,
so appending storage changesets after would fail, as
the cursor is at the *last* storage slot, but we try
to insert a storage slot that is not necessarily the last.

However, we know that if storage was wiped, then all
storage slots changed anyway; so the fix is relatively
simple, as even if it was not a bug, it would be inefficient
to write the same entry twice.