fix: correct storage changesets
In the case that an account is self-destructed and then
re-created in the same block, where the re-creation writes
to a slot that we have never written to before,
we would end up with an incorrect changeset.

The fix is to load storage values from the database
and merge them with the in-memory changeset before
writing.