chore: remove prune_modes from BlockWriter
This removes the `prune_modes` argument from the `BlockWriter` trait,
because the argument leaks a node implementation detail into the storage
abstraction. Instead, the it is included as a field in
`ProviderFactory`, as well as `DatabaseProvider`, although it is only
used in the RW version of `DatabaseProvider`.

When the blockchain tree is initialized, the prune modes are set, so
that we achieve the same behavior as before when writing.