feat: introduce static file task
This implements a basic static file task, which takes a single type of
action containing the most recent finalized block number. It will then
run the static file producer and write to static files.

For integration: when this task sends a response, an action should be
sent to the persistence task.