perf(era): Skip download if ERA file with verified checksum exists
The download `folder` can have files in it that are above the current `index`.
This happens if the node import or sync was interrupted and then restarts.

In that case, we want to use what is already downloaded if its checksum is verified.
Otherwise we delete it and download again.