feat(era): Delete files outside the range before counting them
The ERA stream can be initialized with a directory that already has some ERA files in it.
However, some of those files can be outside of the `index + max_files` range, where
`index` is the currently processed file and `max_files` is the maximum number of ERA files
within `folder`.

The files that are outside of this range should be deleted, because they are in conflict
with the rule of keeping up-to `max_files`. Can even cause the downloader to freeze until
the files are manually removed.