Introduce script to compare reth-bench latency CSVs
This adds a python script to reth-bench which which plots graphs
comparing two combined_latency.csv files output by reth-bench. The
graphs which are plotted are:

- A histogram of the percent change between latencies, bucketed by 1%
  increments.

- A simple line graph plotting the latencies of the two files against
  each other.

The script is optional to use, but makes it quick an easy to compare
profiles. Having a histogram in particular is nice because it takes the
guesswork out of understanding both the median change and how consistent
that change is.