feat(draft): ParallelSparseTrie reveal_node #16863
The objective here was to mostly copy over the existing `reveal_node`
implementation from the SparseTrie. When revealing a path there are two
possible consequences:

- One or more child nodes of a branch/extension where revealed
  recursively. In general there will only be one level of recursion, but
  (I _think_ it might be possible for there to be two recursions if an
  extension points to a branch which has two small leaves.)

- A leaf was revealed (or more than one, if child leaves were revealed,
  see above point). In this case the the ParallelSparseTrie keeps leaf a
  fullpath->value mapping, which needs to be updated.

What added complication was that the ParallelSparseTrie is split into
two tries for any path: the upper and lower trie. This means that if a
path is revealed on the upper path, but has child nodes to be revealed
recursively, those child nodes might be found on one or more lower
subtries. Conversely, when revealing on a lower subtrie the revealed
leaves need to be propagated back to the ParallelSparseTrie.

I attempted to deal with this as best as possible without duplicating a
ton of code. The result is maybe not 100% optimized; there's some Vecs
floating around in the RevealNodeOutcome that I'd rather get rid of
maybe. I think it'd be possible to introduce a Vec pool though, as they
have fixed max capacities.

TODO

- Confirm my approach here makes sense.
- Discuss how parallelization is going to actually work in this case.
- Tests.