perf(trie): Implement get_leaf method on ParallelSparseTrie
This method will be useful for implementing `remove_leaf` (#16869), and
probably will be useful for other methods too. This is necessary because
we don't have a mapping from the full key of a leaf to the path of that
leaf, so we must traverse the trie from root when we want to find it.