Place the root nodes of the lower SparseSubtries in those tries
Previously the root nodes of each lower SparseSubtrie (0x00, 0x01, 0x02,
etc..) would get placed in the upper SparseSubtrie. This is inconvenient
for `update_hashes` and other methods which we are implementing, so this
PR moves those nodes into their respective lower trie instance.

In implementing this I found a couple of couplings between the logic of
the `SparseSubtrieType` and `reveal_node`; a new method is added to
`SparseSubtrieType` to remove this coupling.