chore(trie): factor out SparseTrieState
This change removes the need for the SparseTrieState type, and the
associated `SparseTrie::AllocatedEmpty` variant, by instead passing
around the reset RevealedSparseTrie itself in order to re-use
allocations across payloads.

The `SparseTrie::Blind` variant is modified to now optionally carry a
`RevealedSparseTrie`, and the SparseTrieTask will take that out of the
SparseStateTrie to re-use.

The `RevealedSparseTrie::from_root` method is superceded by the new
`RevealedSparseTrie::with_root`, which performs essentially the same
function but allows us to re-use the RevealedSparseTrie while
maintaining correct semantics around the empty root node.