fix(trie): track branch node updates only in ParallelSparseTrie, not subtries
This commit contains 3 bug fixes related to branch node update
retention.

**Fix 1**

The ParallelSparseTrie previously kept the `Option<SparseTrieUpdates>`
field, used to both enable and collect branch node updates, in multiple
places:

* 1 on the ParallelSparseTrie
* 1 on the upper SparseSubtrie
* 1 on each of the 256 lower SparseSubtries.

This caused issues in `remove_leaf`, which needs to fully remove a lower
subtrie if it no longer has any nodes, and would therefore lose any
accumulated branch node updates.

The fix is to only track branch node updates on the ParallelSparseTrie.
This introduces a new wrinkle, which is that `update_hashes` runs in
parallel on each lower subtrie and needs to make changes to the global
update set. To get around this we track these changes using a
`Vec<SparseTrieUpdatesAction>` on each lower subtrie, and then
apply/drain those changes in `update_subtrie_hashes`.

**Fix 2**

`with_updates` wasn't properly propagating the setting to the lower
subtries. Existing lower subtries wouldn't receive the setting, and any
created afterwards wouldn't either. This has been fixed.

**Fix 3**

Fixing the above revealed that `update_leaf` had a bug which went
unnoticed before because update retention wasn't actually active. When
turning an extension into a branch, it's possible `update_leaf` would
need to reveal the previous child of the extension. This was failing in
cases where the extension was on the upper trie and the child on a lower
one. This has been fixed and a test case added.