fix(trie): properly propagate update retention in ParallelSparseTrie
This commit contains 3 bug fixes related to update retention.

**Fix 1**

The ParallelSparseTrie previously kept the `Option<SparseTrieUpdates>`
field, used to both enable and collect branch node updates, in multiple
places:

* 1 on the ParallelSparseTrie
* 1 on the upper SparseSubtrie
* 1 on each of the 256 lower SparseSubtries.

The one located on the ParallelSparseTrie and upper SparseSubtrie were
essentially redundant; the updates field on the ParallelSparseTrie has
been removed.

**Fix 2**

`with_updates` wasn't properly propagating the setting to the lower
subtries. Existing lower subtries wouldn't receive the setting, and any
created afterwards wouldn't either. This has been fixed.

**Fix 3**

Fixing the above revealed that `update_leaf` had a bug which went
unnoticed before because update retention wasn't actually active. When
turning an extension into a branch, it's possible `update_leaf` would
need to reveal the previous child of the extension. This was failing in
cases where the extension was on the upper trie and the child on a lower
one. This has been fixed and a test case added.