fix(trie): correctly handle path field on cleared ParallelSparseTrie lower subtries
**Background**

The lower subtries of ParallelSparseTries have a `path` field which
indicates the "root node" of the lower subtrie, that is the node with
the shortest path held by the lower sparse trie. This field is used in a
few methods to aid in walking a lower subtrie without first having to
walk the upper.

The `path` was not getting properly reset on `clear()`, resulting in
state leaking between re-use of the ParallelSparseTrie.

**Fix**

For maximum correctness a new enum is introduced, LowerSparseSubtrie,
which is used to track whether or not a lower subtrie is revealed, or
just blinded while re-using a previously allocated `SparseSubtrie`. We
never reveal a LowerSparseSubtrie without also revealing its root path
at the same time.

Another small fix here is to insert an empty node into the root path of
the _upper_ subtrie on `clear()`. This makes the ParallelSparseTrie
semantically consistent with its serial implementation, though should
have very little effect otherwise.