fix(trie): ParallelSparseTrie: remove leaves from upper subtrie when update in a lower
`update_leaf` will always start by inserting the new leaf value into the
upper subtrie's `values` hashmap. It does this to determine if the leaf
already exists in the upper subtrie, and short-circuits the call there
if so.

There were some edge-cases where `update_leaf` would not later remove
that from its `values` hashmap, even after determining that the leaf
belongs to a lower subtrie. This fixes that.

This also contains a full test case based on block 1 of the hoodi
network, which is what originally exposed this issue.