chore(trie): Implement ParallelSparseTrie::find_leaf
This implements the final method of the SparseTrieInterface trait. We
don't need this method in the immediate term, but it'd be better to
round out the trait rather than have someone get rugged by it not being
there later.

The implementation is able to take advantage of `find_next_to_leaf`, and
so is fairly simple. The tests are all taken from the RevealedSparseTrie
more or less verbatim.