perf(trie): parallelize ParallelSparseTrie::reveal_nodes
(Proof-of-concept, WIP, draft, TODO, do not merge)

Introduce `reveal_nodes` to the ParallelSparseTrie, which first sorts
the provided nodes so that they can then be grouped by their subtrie and
revealed in parallel.

We are currently sorting the nodes as part of SparseStateTrie prior to
revealing. The idea is that we would stop sorting there, and instead
sort within SerialSparseTrie and ParallelSparseTrie. Each needs a
different sorting algorithm, so it makes more sense for them to sort on
their own terms.