perf(trie): Re-use storage tries across payloads
This is implemented by re-using the entire SparseStateTrie across
payloads, so that we keep reusing the accounts trie as well.

There's a small change to how storage tries are fetched/created in
SparseStateTrie, the rest of this change has to do with passing
SparseStateTrie around. A new type is introduced for that,
ClearedSparseStateTrie, which makes it easy to enforce at compile-time
that only a cleared state is being used in a payload task.

Fixes #17457