Track added keys in multiproof generation
In #17643 we introduced tracking of removed keys within a block in order
to fully reveal sparse tries prior to updating/removing leafs. This lets
us never have to block root calculation to reveal blinded nodes in the
sparse tries.

This was implemented by assuming that all keys are added (as opposed to
being only modified), which results in over-revealing the sparse tries
with all extension node children. This isn't logically incorrect, but is
a bit wasteful.

This change implements proper tracking of added keys alongside tracking
of removed keys. When a key is added we always generate its proof in
`on_state_update`, just like key removal. Because we can now enforce
that added keys have their proofs generated in `on_state_update` we no
longer need to optimistically retain extra proofs in prewarms.