fix(trie): Fix call to update_account in witness
In 8193fcff9 (#17643) the behavior of `update_account` was changed so
that it does not call `remove_account_leaf` on its own, but rather
returns a boolean indicating that the caller should do so. This change
in behavior was not accounted for in the witness code which also calls
this method.