fix: Pass prefix set from init_from_state_dump into compute_state_root
init_from_state_dump is intended to be usable when there is already
state data in the db, after genesis. However if we do not pass the
prefix sets of changed accounts/storage slots into the state root
computation then the trie db tables are not populated correctly and we
get a state root mismatch.