chore(trie): dont warn on blinded node reveals
For now these are still expected, there seem to be some edge-cases that
are still not handled. It's not worthwhile to scare uses with WARNs for
these, as they do not indicate anything is actually wrong, the sparse
tries will simply fallback to the behavior they already have in stable.