fix(trie): Update sync stage checkpoint in repair-trie
When we run `repair-trie`, at the end of the run we can be certain that
the trie tables are consistent with the HashedAccounts/Storages tables,
and so we can update the sync checkpoint for it accordingly.

Some assertions are added on the HashedAccounts/Storages checkpoints
here in order to check that those two tables are consistent with each
other.

This handles the case of someone partially completing a merkle sync
pipeline and then subsequently running repair-trie.