feat: Write trie changesets to DB on engine persistence
Closes #18465

This introduces two new provider methods for writing the trie changeset
data based on the current Accounts/StorageTrie dataset. These will be
used in the engine API (as added by this PR) as well as in the pipeline
sync.

There is also a slight refactor; I added a helper for generating the
storage changesets in the same module space as the existing helper for
storage reverts, which was alone in a module called `bundle_state`. I
renamed the module to `changeset_utils` since everything in side has to
do with populating changeset tables.