chore(trie): Overlay arguments for write trie changeset methods
This allows for passing optional overlays into the
`write_trie_changesets` and `write_storage_trie_changesets` provider
methods.

The overlay is a TrieUpdates which is used to augment the state of the
trie db tables. Using the overlay we can write changesets as if the DB
is at a previous block, which will be used during pipeline sync.

Implementing this change required refactoring the
StorageTrieCurrentValuesIter utility to accept a TrieCursor rather than
a normal DbCursor. It also required implementing a TrieCursorIter which
wraps a TrieCursor into an Iterator, for passing in to
`storage_trie_wiped_changeset_iter`. Using both of these changes we
could use an InMemoryTrieCursor instead of a direct db cursor.