fix(trie): Rewrite InMemoryTrieOverlay (with proptests!)
The `InMemoryTrieOverlay::next` method was previously very broken, in
that it did not ever descend into lower parts of the trie. For example
a given current key of `0xA`, and subsequent keys of `0xA0` and `0xB`,
`next` was returning `0xB`. This was due to incorrect usage of the
`Nibbles::increment` method.

Fixing this issue involved essentially rewriting the entire type. The
new implementation includes proptests to ensure that it functions
correctly.

Some extra debug checks have also been added to ensure that the overlay
is actually being used correctly.