feat(exex): Provide TrieUpdates in all ExEx notifications
This changes the TrieUpdates field of the Chain type (which is used in
ExEx notifications) from an Option into a BTreeMap mapping the block
number to the updates pulled from the ExecutedBlock. With this in place
all ExEx subscribes have full access to all trie updates produced by
each block.