feat: fix metrics log line when passed metrics port is 0 (#334)
Fixes #333

The simplest fix is moving the `log` to when the port binding happens
i.e. inside `start_endpoint`.
The only way to get the `port` is by binding to a listener which happens
in `start_endpoint`. If we bind before we will get another value later
(bind will produce different values if 0 is used)