perf: box ForkId in Peer struct to reduce size
ForkId contains a ForkHash (4 bytes) and a u64 (8 bytes), but the field
is optional and not always present. Boxing it reduces the size of the Peer
struct by moving the allocation off the stack when present, improving memory
layout for the common case where many peers don't have fork_id set.