chore(trie): Allow reusing Hashed/TrieCursors
Towards #19512

In order to implement a reusable proof calculator we first need the
underlying cursors used by the calculator to be reusable.

For account cursors this isn't so difficult, a new `reset` method is
introduced which resets the ForwardInMemoryCursor and other state
fields.

For storage cursors the situation is complicated because reusing the
cursor might involve using it for a completely different hashed address.

To handle this the storage cursors are given a `set_hashed_address`
method which effectively resets them, as well as pulls out the correct
overlay for the chosen address. Implementing this requires that the
cursors now hold onto the full
`&TrieUpdatesSorted`/`&HashedPostStateSorted`.

It also requires slightly
different handling of the wiped case; before we were simply storing an
underlying cursor in an Option, with None indicating wiped, but now we
need to always have an underlying cursors (so we can reuse it for a
future non-wiped storage). A new boolean is introduced instead.