feat(trie): Implement set_hashed_address on mock trie cursors
This properly implements the `set_hashed_address` method on the
MockTrieCursor and MockHashedCursor. This method was previously left
unimplemented, but will be necessary for an upcoming feature.

Implementing this method required changing the internals of the mock
cursors a bit, such that they are able to hold onto the full set of
storage tries in order to switch between them, similar to how the
in-memory cursor overlays do.