feat(cli): repair-trie metrics
This adds an optional metrics endpoint to the `reth db repair-trie`
sub-command, which can be used to track in prometheus when
inconsistencies are found, which makes alerting much more reliable.

This also adds a small log message to include the current db tip, which
is convenient to know when debugging issues.